소스코드/flask/requirements.txt
aniso8601==9.0.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chromedriver-autoinstaller==0.6.2
click==8.1.7
cryptography==41.0.4
django-environ==0.11.2
exceptiongroup==1.1.3
Faker==19.6.2
Flask==2.3.3
flask-restx==1.1.0
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
oracledb==1.4.1
outcome==1.2.0
packaging==23.1
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
webdriver-manager==4.0.0
webencodings==0.5.1
Werkzeug==2.3.7
wsproto==1.2.0
zipp==3.16.2
